common/input
